refactored StepSolve things into only console
Forwad sim command goes until next 'edge'
haven't tested in very many configurations yet

+++ console/ConsoleTimer.cpp
@@ -2,8 +2,8 @@
 #include "ConsoleTimer.h"
 #include <chrono>
 
-unsigned long long ConsoleTimer::programTimeMs() {
-  return std::chrono::duration_cast<std::chrono::milliseconds>(
+unsigned long ConsoleTimer::programTimeMs() {
+  return (unsigned  long)std::chrono::duration_cast<std::chrono::milliseconds>(
       std::chrono::steady_clock::now().time_since_epoch()).count();
 }
 #endif
