+++ lib/commanduino/TimerInterface.h
@@ -3,5 +3,5 @@
 class TimerInterface {
 
   /// \brief the time since the start of the program in milliseconds
-  public: virtual unsigned long long programTimeMs() = 0;
+  public: virtual unsigned long programTimeMs() = 0;
 };
