+++ sim/SimTimer.h
@@ -4,10 +4,10 @@
 
 class SimTimer : public TimerInterface {
   public:
-    virtual unsigned long long programTimeMs() override;
+    virtual unsigned long programTimeMs() override;
     static void simTimeCallback(ConstWorldStatisticsPtr &msg);
 
   private:
-    static unsigned long long simTimeMs;
+    static unsigned long simTimeMs;
 
 };
